% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edf.R
\name{edf}
\alias{edf}
\title{edf function}
\usage{
edf(x, X)
}
\arguments{
\item{x}{numeric, function argument}

\item{X}{vector, empirical data}
}
\value{
numeric, left continuous empirical distribution function value
}
\description{
Computes the value of the left continuous empirical distribution function at a given point.
}
\examples{
edf(x = 0, X = c(- 1, 0, 1));
}
