% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edf.R
\name{edf}
\alias{edf}
\title{A edf function}
\usage{
edf(x, X)
}
\arguments{
\item{x:}{function argument as numeric}

\item{X:}{empirical data as vector}
}
\description{
This function calculates value of left continuous empirical distribution function at given point
}
\examples{
edf(x = 0, X = c(-1,0,1));
}
\keyword{distribution}
\keyword{empirical}
\keyword{function}

